third_party/WebKit/LayoutTests/external/wpt/fullscreen/api/document-exit-fullscreen-nested-in-iframe-manual-expected.txt

This is a testharness.js-based test.
FAIL Document#exitFullscreen() for nested fullscreen inside an iframe assert_equals: expected Element node <iframe allowfullscreen="" srcdoc="<div><div></div></div>... but got null
Harness: the test ran to completion.
